Move stroke to own file.
